refactor: Refactor fetching Egg Inc contract data handling and updating logic
- Moved logic to get the EI contract data out of the main.go source into its own sources
